+++ kong/plugins/opentelemetry/handler.lua
@@ -37,6 +37,11 @@ local default_headers = {
 local queues = {} -- one queue per unique plugin config
 local headers_cache = setmetatable({}, { __mode = "k" })
 
+-- we use plugin instance ID as the queue ID
+local function get_queue_id()
+  return kong.plugin.get_id()
+end
+
 local function get_cached_headers(conf_headers)
   if not conf_headers then
     return default_headers
@@ -153,7 +158,7 @@ end
 function OpenTelemetryHandler:log(conf)
   ngx_log(ngx_DEBUG, _log_prefix, "total spans in current request: ", ngx.ctx.KONG_SPANS and #ngx.ctx.KONG_SPANS)
 
-  local queue_id = conf.__key__
+  local queue_id = get_queue_id()
   local q = queues[queue_id]
   if not q then
     local process = function(entries)